UI: Refactor types and change IHologram fields
