+++ HoloRepositoryUI/src/types/AppInteractions.tsx
@@ -1,0 +1,4 @@
+export enum HologramCreationMode {
+  generateFromImagingStudy = "GENERATE_FROM_IMAGING_STUDY",
+  uploadExistingModel = "UPLOAD_EXISTING_MODEL"
+}
